Add reminder to delete more resources
As pointed out on the documentation of the [DeleteUser method](https://docs.aws.amazon.com/sdk-for-go/api/service/iam/#IAM.DeleteUser)
"you must delete the items attached to the user manually, or the deletion fails".

Currently we are simply creating, then deleting, a user but not
associating any Role, credentials or passwords to it.