This allows IAM Service Linked roles to be deleted.
In the Gruntwork Reference Architecture, we need to clean up AWSServiceRoleForAutoScaling.

Note that this is a naive approach; it requests the resource be deleted, but doesn't verify
that they are actually deleted.

We could do this verification by capturing the DeletionTaskId and then querying for the deletion
status until the status is (hopefully) SUCCEEDED. Alternate statuses are NOT_STARTED, IN_PROGRESS, and FAILED.