Move telemetry down in README (#658)
I've found it very annoying that the very top of the cloud-nuke README file is a note about telemetry. Being up front about this is important, but it shouldn't appear before the README even introduces what the tool is or does!

This PR moves the telemetry note below that, but still above install, so it should still get high visibility.